backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
newspaper3k==0.2.8
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
lxml==4.9.3
hypothesis==6.92.0
pytest==7.4.3
